requirements.txt
Django==1.11.25
django-compressor==1.3
dj-database-url==0.2.2
gunicorn==19.1.0
Fabric==1.6.1
django-autoslug==1.7.2
sorl-thumbnail==12.5.0
pixelfields-smart-selects==0.1.5
-e git://github.com/javierwilson/django-chartit.git#egg=django_chartit
-e git://github.com/javierwilson/django-model-report.git@930f2b2a793cdb1bafd8927100d4aaacfc299721#egg=django_model_report
-e git://github.com/jakewins/django-money.git@fe0b20831a15cef3f5225392276fdd4737de216f#egg=django_money
django-mathfilters==0.3.0
django-lockdown==2.0.0
django-debug-toolbar==1.4
django-smart-selects==1.2.9
